requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
distlib==0.3.4
Django==4.0.1
django-jazzmin==2.4.9
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.0
jmespath==0.10.0
phonenumbers==8.12.41
Pillow==9.0.0
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.10.0
virtualenv-clone==0.5.7